MD: move MD administration towards KS
