+++ Services/MetaData/classes/Settings/class.ilObjMDSettings.php
@@ -1,0 +1,31 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * This file is part of ILIAS, a powerful learning management system
+ * published by ILIAS open source e-Learning e.V.
+ *
+ * ILIAS is licensed with the GPL-3.0,
+ * see https://www.gnu.org/licenses/gpl-3.0.en.html
+ * You should have received a copy of said license along with the
+ * source code, too.
+ *
+ * If this is not the case or you just want to try ILIAS, you'll find
+ * us at:
+ * https://www.ilias.de
+ * https://github.com/ILIAS-eLearning
+ *
+ *********************************************************************/
+
+/**
+ * @author  Stefan Meyer <meyer@leifos.com>
+ */
+class ilObjMDSettings extends ilObject
+{
+    public function __construct($a_id = 0, $a_call_by_reference = true)
+    {
+        $this->type = "mds";
+        parent::__construct($a_id, $a_call_by_reference);
+    }
+}
