--- Services/MetaData/templates/default/tpl.settings.html
@@ -1,3 +1,0 @@
-{SETTINGS_TABLE}
-<br />
-{COPYRIGHT_TABLE}
