+++ Services/MetaData/templates/default/tpl.show_copyright_row.html
@@ -26,14 +26,10 @@
 	<td class="std">
 		{VAL_STATUS}
 	</td>
-	<!-- BEGIN link_bl -->
+	<!-- BEGIN actions -->
 	<td class="std">
-		<a class="il_ContainerItemCommand" href="{EDIT_LINK}">{TXT_EDIT}</a>
-		<!-- BEGIN link_usage -->
-		<br />
-		<a class="il_ContainerItemCommand" href="{USAGE_LINK}">{TXT_USAGE}</a>
-		<!-- END link_usage -->
+		{ACTIONS}
 	</td>
-	<!-- END link_bl -->
+	<!-- END actions -->
 </tr>
